Moved all files  utils into their own folder
