+++ src/commands/executeCommand.ts
@@ -2,7 +2,7 @@ import { Message } from 'discord.js'
 import TimeAgo from 'javascript-time-ago'
 import en from 'javascript-time-ago/locale/en'
 import { logger } from '../logger'
-import { findOrMakeGuild, parseMessage } from '../utils'
+import { findOrMakeGuild, parseMessage } from '../utils/utils'
 import { attack } from './attack'
 import { defend } from './defend'
 import { describeMonster } from './describe'
